pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spreadsheet-qa"
version = "0.1.0"
description = "Tablerreur — Spreadsheet QA, audit, validation and export desktop tool"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6",
    "pandas>=2.1",
    "openpyxl>=3.1",
    "pyyaml>=6.0",
    "rapidfuzz>=3.6",
    "chardet>=5.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "pytest-qt>=4.4",
]

[project.scripts]
spreadsheet-qa = "spreadsheet_qa.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/spreadsheet_qa"]
# Ensure YAML template files are included in the wheel
include = ["src/spreadsheet_qa/resources/**/*.yml", "src/spreadsheet_qa/resources/**/__init__.py"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-x -q"
